requirements.txt
# Core Dependencies
fastapi==0.95.0
uvicorn==0.21.1
python-dotenv==1.0.0
pydantic==1.10.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.12
alembic==1.10.4
psycopg2-binary==2.9.6
redis==4.5.5
python-dateutil==2.8.2
requests==2.28.2

# AI/ML Dependencies
transformers==4.28.1
torch==2.0.0
numpy==1.24.3
pandas==2.0.1
scikit-learn==1.2.2

# Development Dependencies
pytest==7.3.1
pytest-cov==4.0.0
black==23.3.0
isort==5.12.0
mypy==1.2.0
flake8==6.0.0

# Documentation
mkdocs==1.4.2
mkdocs-material==9.1.8

# Security
cryptography==40.0.2
python-json-logger==2.0.6

# Email
aiosmtplib==2.0.2
email-validator==1.3.1

# Caching
redis-py==4.5.5

# Async
httpx==0.24.0
anyio==3.6.2

# Utils
python-slugify==8.0.1
python-magic==0.4.27
Pillow==9.5.0
